Implemented Rule-of-Zero on SensorInterface
Since `ros2_ouster::SensorInterface` is a base class and intended to be derived
from, we need a virtual dtor. To that end, to be in compliance with the
Rule-of-Zero we need to provide explicit implementations for "the other four"
special functions (those related to copy and move semantics). Also note, we
want to avoid "slicing" (in the future ... maybe) so we `delete` the copy
operations rather than `default` them.

See: https://en.cppreference.com/w/cpp/language/rule_of_three
(including the referenced CPP Core Guidlines it links to in the "Rule of
Zero" section)