Added runtime-selectable QoS support to allow for recording sensor data via rosbag2 in Eloquent.
Apparently, as of this writing, rosbag2 cannot record topic data whose QoS has
a reliability level of anything but `RELIABLE`. See:
https://github.com/ros2/rosbag2/issues/125. The `rmw_qos_profile_sensor_data`
(which we are using in this project *a priori*) employs `BEST_EFFORT`
reliability. I've added a new parameter
`use_system_default_qos_for_sensor_data` to the node and parameter file. By
default this is set to `False` which maintains the original behavior of this
package. However, if set to `True`, rather than `rmw_qos_profile_sensor_data`
we use ``rmw_qos_profile_default` as our QoS profile for the sensor
topics. This (supposedly) most closely mimics ROS (classic) behavior but more
importantly, allows rosbag2 to record data from these topics. Hopefully this
issue with rosbag2 gets addressed by Foxy and we can deprecate this
feature. But for now, not having the ability to record data for offline
analysis is very limiting and we need something like this in this driver
package.