Reduced jitter in the pointcloud_processor publisher
The driver now gives up ownership of the PointCloud message in the
pointcloud_processor at publish time to allow for middleware optimizations to
include the elimination of unnecessary copies. Several tests on my machine show
order of magnitude jitter improvements through the driver + middleware stack
both out-of-process and while running as a component to sub-millisecond
levels.