Buffered packet processing (#97)
* Simplify lidar scan assembly code

As double-buffering is used for the scan assemnly (in contrast to the ouster example)
we can simply zero out our buffer for the new frame. We then no longer need to
special-case zeroing-out missing packets.

* Initialize FullRotationAccumulator with batchReady == false

* Implement buffered packet processing

This commit splits the processing of lidar data into two seperate
threads. Previously, the code receiving sensor data from the wire
was blocked by the lidar or imu processors, causing dataloss on
slower systems (Issue #96).

* Modifies sensor interface so receiving functions get passed a
  pointer to a buffer to write their data to
* Implements ringbuffers for lidar and imu data in ouster_driver
* Splits processing into two seperate threads:
  * receiveData receives from the wire and fills ringbuffers
  * processData takes packets from the queue, assembles the lidar
    scan and runs the processors
* Threads are implemented as pure threads instead of ros timers,
  as timing inconsistencies for the receiver thread were also
  leading to data loss

* Revert "Simplify lidar scan assembly code"

This reverts commit 581b35f14ef098041f03131deb18094ea92b98df.
This modification to the client code is not necessary for the
buffered packet processing, and conflicts with changes since
made in the ouster_examples repository.

* Lock against race on [lidar|imu]_packet_head

Moves the modification of the head indices to before the mutex
unlock. This is only necessary for the warning about buffer
overruns to work correctly. We don't need to lock against the
actual data in the buffer.

* Move ringbuffer code to it's own header

* Fix linting issues

Co-authored-by: Mirko Kugelmeier <mirko.kugelmeier@rwth-aachen.de>